fix(dependencies): Do not invoke servers in their modules
It turns out that Invoke functions in modules run before Invoke functions
in the parent. This means that the http and grpc servers will still start
before their services are started.

With this in mind, I thought it was best if users explicitly call the
Invoke functions in their apps: this gives them total control over the
ordering and removes the significance of the order in which modules are
embedded in the system.

This behaviour will have to be documented.