fix(logging): Fixes a racy test
The test is racy because the assert happens in the "client" goroutine,
which runs concurrently from the server handler goroutine that is the actual
system under test.
By adding the extra recv, we get the io.EOF error: this is the server
informing us its done. This should mean the server goroutine is finished
and we definitely have the log.