feat(fxgrpc): Add the ability to provide named grpc servers
This is implemented by providing an option on the module constructor.
By making use of the recent fx.Private functionality, we only have to
annotate the "output" of the module and not the internals.

The health and reflection modules will only work with the standard
module output. They will not install their services on named servers.