feat(fx): Hide module conf with fx.Private
The latest version of fx allows you to mark values inserted with
Supply to be marked as private.
This has some benefits in our modules and allows us to remove
name annotations.
I also restructured the fxhttp named module to work like the grpc
one, where we use a vararg on the normal constructor.

In a subsequent PR I want to try and remove more name annotations.