fix(cert-reloader): Use time based strategy instead of watching fs events
Correctly watching for file system events is extremely tricky on linux.
In fact it is probably not possible to correctly watch for events in the
general case if the file being watched is a symbolic link.

Unfortunately kubernetes secret mounts are implemented via symbolic
links, leading to production issues where kube workloads would not
correctly refresh their TLS certificates.

After some experimentation with trying to resolve symbolic links
and additionally watching the underlying paths, we decided that the
implementation was becoming much too complex and abandoned the approach.

The new implementation uses a simple ticket on a goroutine to reload
the data from disk. If reloading fails, we return the last succesfully
loaded data. The interval between reloads is configurable, and a default
of 1h is provided, which should work for most certificate cases.

Even for intervals as short as 1m, the overhead of the reloading should
be impossible to measure on any program.